Add two static pages: experiments and toremember
Add two static pages. The first one is called experiments in which I'm
going to set links to some experiments that I'm going to be doing while
learning new technologies. The second one is a page which will contain
raw data and raw thoughts and ideas that I may be need to remember in
the future. It is like my daily txt file that I used to keep track of
everything I need to do. I've tested many apps and in the end what suits
me best is just a simple txt file.

Contact info: joaquin@isla.io